+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.2.0] - 2019-09-14
+### Added
+- `MTL`-style type class for logging with instance that logs to `fast-logger`.
+- `sourceLoc` field to `Log` record type for logging of location in the source code.
+
 ## [0.1.0.1] - 2018-09-22
 ### Chaged
 - `Text` for the messages and applications replaced with more abstract class.